move load_and_run_plugins, unregister plugin NRR after loading
