+++ Cargo.toml
@@ -26,7 +26,7 @@ orbits = { git = "https://github.com/blu-dev/orbits" }
 smash-arc = { git = "https://github.com/jam1garner/smash-arc", features = ["smash-runtime", "rust-zstd"] }
 
 [patch.crates-io]
-ring = { git = "https://github.com/skyline-rs/ring", rev = "9b5eaee" }
+ring = { git = "https://github.com/skyline-rs/ring", branch = "0.16.20" }
 webpki = { git = "https://github.com/skyline-rs/webpki" }
 
 [patch.'https://github.com/ultimate-research/nnsdk-rs']
@@ -35,8 +35,8 @@ nnsdk = { git = "https://github.com/blu-dev/nnsdk-rs", branch = "patch-4" }
 [patch.'https://github.com/jam1garner/smash-arc']
 smash-arc = { git = 'https://github.com/blu-dev/smash-arc' }
 
-[patch.'https://github.com/blu-dev/orbits']
-orbits = { path = "../../orbits" }
+# [patch.'https://github.com/blu-dev/orbits']
+# orbits = { path = "../../orbits" }
 
 [profile.dev]
 panic = "abort"
